requirements.txt
pyyaml==6.0.1
requests==2.31.0
aiohttp==3.9.1